k1c: Use $r6 for syscall number
Summary:
In order to handle syscall restart mecanism after handling signal,
we must have a way in the kernel to modify the syscall number and
reexecute the scall bundle. The kernel was already using $r6 and
expected it to be the syscall number. Fix that in syscall.h to
always pass syscall number in $r6.

Ref T6696

Test Plan: Wait for test in initramfs to be ready to test this specific feature.

Reviewers: O52 uClibc Coolidge, mpoulhies

Reviewed By: O52 uClibc Coolidge, mpoulhies

Subscribers: mpoulhies

Maniphest Tasks: T6696

Differential Revision: https://phab.kalray.eu/D512