kvx: Fix stack usage in clone.S
Summary:
Various problems have been spotted in __clone:
- Stack is used before being reserved
- Stack alignment is not enforced

In order to fix these, first thing to do is to align the child stack and
then increment it to store arguments. This also requires to stay aligned
on a 32 byte boundary and thus increment the stack by 32 bytes even if
only 16 are needed. Moreover, child stack is now reserved in __clone and
not in child_start which is wrong. Indeed, if a signal is delivered to
the child before the process started, the child stack can potentially be
overwritten by the signal frame creation.

Ref T13064

Test Plan:
Verify that tests are passing and more particularly tst-eintr1 which
raised this specific problem.

Reviewers: O52 Linux libcs Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T13064

Differential Revision: https://phab.kalray.eu/D5883